Recover automatic block number calculation
The total block count was once done automatically but it was lost on commit
457078c90d5a3738ee2ed2206728b1b5cacecb9f.

The calculation is recovered, but made optional: it is only done if total blocks
is not passed by cmdline. If passed, that value is used and not touched.

Also, reserved_frac is taken into account on the automatic calculation.