Enable system locale - change from the standard (C) to system locale.
This allows libarchive (in case it is activated) to handle filenames.
We only change LC_CTYPE since libarchive only needs the charset set.
We don't use LC_ALL because it causes problems on some systems.
We restore the original LC_CTYPE after extraction to avoid side effects.
We use uselocale instead of setlocale to avoid setting LC_CTYPE globally.
See on libarchive Website for a more complete description of the issue:
 https://github.com/libarchive/libarchive/issues/587
 https://github.com/libarchive/libarchive/wiki/Filenames
 https://github.com/sbabic/swupdate/blob/master/handlers/archive_handler.c#L94